apps/api/pyproject.toml
[project]
name = "wsp-agent-api"
version = "0.1.0"
description = "WhatsApp agent backend (Twilio + FastAPI)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.115.0",
  "prometheus-client>=0.21.0",
  "psycopg[binary]>=3.2.0",
  "slowapi>=0.1.9",
  "python-multipart>=0.0.9",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy[asyncio]>=2.0.36",
  "asyncpg>=0.30.0",
  "redis>=5.0.0",
  "pydantic-settings>=2.6.0",
  "twilio>=9.3.0",
  "openai>=1.55.0",
  "google-genai>=1.0.0",
  "google-api-core>=2.19.0",
  "google-api-python-client>=2.150.0",
  "google-auth>=2.35.0",
  "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.8.0",
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=75.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
